+++ backend/src/main/java/com/techeer/carpool/domain/auth/dto/AuthTokens.java
@@ -1,3 +1,3 @@
-package com.techeer.carpool.domain.member.dto;
+package com.techeer.carpool.domain.auth.dto;
 
 public record AuthTokens(String accessToken, String refreshToken) {}
