+++ backend/src/main/java/com/techeer/carpool/domain/auth/dto/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.techeer.carpool.domain.member.dto;
+package com.techeer.carpool.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
