+++ backend/src/main/java/com/techeer/carpool/domain/auth/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
-package com.techeer.carpool.domain.member.repository;
+package com.techeer.carpool.domain.auth.repository;
 
-import com.techeer.carpool.domain.member.entity.RefreshToken;
+import com.techeer.carpool.domain.auth.entity.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
