+++ backend/src/main/java/com/techeer/carpool/domain/auth/service/MemberLoginService.java
@@ -1,11 +1,11 @@
-package com.techeer.carpool.domain.member.service;
+package com.techeer.carpool.domain.auth.service;
 
-import com.techeer.carpool.domain.member.dto.AuthTokens;
-import com.techeer.carpool.domain.member.dto.LoginRequest;
+import com.techeer.carpool.domain.auth.dto.AuthTokens;
+import com.techeer.carpool.domain.auth.dto.LoginRequest;
+import com.techeer.carpool.domain.auth.entity.RefreshToken;
+import com.techeer.carpool.domain.auth.repository.RefreshTokenRepository;
 import com.techeer.carpool.domain.member.entity.Member;
-import com.techeer.carpool.domain.member.entity.RefreshToken;
 import com.techeer.carpool.domain.member.repository.MemberRepository;
-import com.techeer.carpool.domain.member.repository.RefreshTokenRepository;
 import com.techeer.carpool.global.exception.CarpoolException;
 import com.techeer.carpool.global.exception.ErrorCode;
 import com.techeer.carpool.global.jwt.JwtTokenProvider;
