+++ backend/src/main/java/com/techeer/carpool/domain/member/service/MemberWithdrawService.java
@@ -1,0 +1,27 @@
+package com.techeer.carpool.domain.member.service;
+
+import com.techeer.carpool.domain.auth.repository.RefreshTokenRepository;
+import com.techeer.carpool.domain.member.entity.Member;
+import com.techeer.carpool.domain.member.repository.MemberRepository;
+import com.techeer.carpool.global.exception.CarpoolException;
+import com.techeer.carpool.global.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class MemberWithdrawService {
+
+    private final MemberRepository memberRepository;
+    private final RefreshTokenRepository refreshTokenRepository;
+
+    @Transactional
+    public void withdraw(Long memberId) {
+        Member member = memberRepository.findByIdAndDeletedFalse(memberId)
+                .orElseThrow(() -> new CarpoolException(ErrorCode.MEMBER_NOT_FOUND));
+
+        member.withdraw();
+        refreshTokenRepository.deleteByMemberId(memberId);
+    }
+}
